+++ src/lightly_train/_commands/train_helpers.py
@@ -8,9 +8,9 @@
 from __future__ import annotations
 
 import logging
+import math
 from pathlib import Path
 from typing import Any, Literal, Sized
-import math
 
 from pytorch_lightning import Callback, Trainer
 from pytorch_lightning.accelerators.accelerator import Accelerator
@@ -350,20 +350,14 @@ def get_epochs(
     )
 
     if epochs != "auto":
-        logger.debug(
-            f"Using provided epochs {epochs}."
-        )
+        logger.debug(f"Using provided epochs {epochs}.")
         return epochs
 
     if method_args_cls.default_epochs is not None:
-        logger.debug(
-            f"Using default epochs {method_args_cls.default_epochs}."
-        )
+        logger.debug(f"Using default epochs {method_args_cls.default_epochs}.")
         return method_args_cls.default_epochs
     elif method_args_cls.default_steps is not None:
-        logger.debug(
-            f"Using default steps {method_args_cls.default_steps}."
-        )
+        logger.debug(f"Using default steps {method_args_cls.default_steps}.")
         # Calculate epochs from steps.
         epochs = math.ceil(method_args_cls.default_steps * batch_size / dataset_size)
         logger.debug(